[EM-1668] move ClientState to impl
