+++ stonex-cms-amqp/stonex-cms-amqp-lib/stonex-cms-amqp-cpp/ClientState.cpp
