+++ stonex-cms-amqp/stonex-cms-amqp-lib/stonex-cms-amqp-cpp/ConnectionContext.h
@@ -20,7 +20,7 @@
 #pragma once
 #include <proton/work_queue.hpp>
 #include <proton/connection.hpp>
-#include "../API/ClientState.h"
+#include "ClientState.h"
 
 //AMQP_DEFINES
