+++ stonex-cms-amqp/stonex-cms-amqp-lib/stonex-cms-amqp-cpp/ConsumerContext.h
@@ -23,7 +23,7 @@
 #include <proton/receiver.hpp>
 #include <proton/receiver_options.hpp>
 #include <cms/Destination.h>
-#include "../API/ClientState.h"
+#include "ClientState.h"
 
 namespace cms
 {
