+++ stonex-cms-amqp/stonex-cms-amqp-lib/stonex-cms-amqp-cpp/SessionContext.h
@@ -20,7 +20,7 @@
 #include <proton/work_queue.hpp>
 #include <proton/session.hpp>
 #include <proton/connection.hpp>
-#include "../API/ClientState.h"
+#include "ClientState.h"
 
 namespace cms
 {
